+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Allow custom error on forms without model. [@victorperez](https://github.com/victorperez)
 * Do not support Ruby < 2.3 anymore. [@gssbzn](https://github.com/gssbzn)
 * Add color input type. [@gssbzn](https://github.com/gssbzn)
+* Set multiple attribute for grouped selects also. [@ollym](https://github.com/ollym)
 
 ### Bug fix
 * Improve disabled option to input_field. [@betelgeuse](https://github.com/betelgeuse)